fix(hub): use BIGINT cast for ms timestamps in auto-close queries
CAST(recorded_at AS INTEGER) overflows on Postgres (32-bit) for 13-digit
millisecond timestamps; switch to BIGINT to match the existing
payment-repo/stats-repo convention and work on both SQLite and Postgres.

+++ apps/hub/src/db/repos/channel-repo.ts
@@ -101,7 +101,7 @@ export class ChannelRepo {
       `SELECT c.* FROM channels c
        WHERE c.status = 'open'
          AND COALESCE(
-           (SELECT MAX(CAST(s.recorded_at AS INTEGER))
+           (SELECT MAX(CAST(s.recorded_at AS BIGINT))
             FROM signed_states s WHERE s.channel_id = c.id),
            0
          ) < ?`,
